Refactor OrdersListViewModel to parse the Orders list data
